Integrate daemon with Wpantund
This patch integrates the main application with wpantund to get
openthread information (network name, channel, pan id, xpan id, and
masterkey). The integration is done through DBus API.

Two methods are used to get these properties: Status and GetProp.
The second one is used to get the masterkey and the first one get the
others properties.

Signed-off-by: Vitor Barros <vba@cesar.org.br>